+++ RELEASES.md
@@ -12,6 +12,7 @@ Version 0.8.17 (2022-XX-XX)
   * Border turns grey while overriding steering
   * Added button to bookmark events while driving; view them later in comma connect
 * AGNOS 6
+* Kia Sportage Hybrid 2023 support thanks to sunnyhaibin!
 
 Version 0.8.16 (2022-08-26)
 ========================